+++ src/main/java/com/tsp/new_tsp_front/api/production/service/FrontProductionJpaApiService.java
@@ -3,12 +3,16 @@ package com.tsp.new_tsp_front.api.production.service;
 import com.tsp.new_tsp_front.api.production.domain.FrontProductionDTO;
 import com.tsp.new_tsp_front.api.production.domain.FrontProductionEntity;
 import com.tsp.new_tsp_front.api.production.service.impl.FrontProductionJpaRepository;
+import com.tsp.new_tsp_front.exception.TspException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Map;
+
+import static com.tsp.new_tsp_front.exception.ApiExceptionType.NOT_FOUND_PRODUCTION;
+import static com.tsp.new_tsp_front.exception.ApiExceptionType.NOT_FOUND_PRODUCTION_LIST;
 
 @Service
 @RequiredArgsConstructor
@@ -26,8 +30,12 @@ public class FrontProductionJpaApiService {
 	 *
 	 */
 	@Transactional(readOnly = true)
-	public List<FrontProductionDTO> getProductionList(Map<String, Object> productionMap) {
-		return frontProductionJpaRepository.getProductionList(productionMap);
+	public List<FrontProductionDTO> getProductionList(Map<String, Object> productionMap) throws TspException {
+		try {
+			return frontProductionJpaRepository.getProductionList(productionMap);
+		} catch (Exception e) {
+			throw new TspException(NOT_FOUND_PRODUCTION_LIST, e);
+		}
 	}
 
 	/**
@@ -41,7 +49,11 @@ public class FrontProductionJpaApiService {
 	 *
 	 */
 	@Transactional(readOnly = true)
-	public FrontProductionDTO getProductionInfo(FrontProductionEntity frontProductionEntity) {
-		return frontProductionJpaRepository.getProductionInfo(frontProductionEntity);
+	public FrontProductionDTO getProductionInfo(FrontProductionEntity frontProductionEntity) throws TspException {
+		try {
+			return frontProductionJpaRepository.getProductionInfo(frontProductionEntity);
+		} catch (Exception e) {
+			throw new TspException(NOT_FOUND_PRODUCTION, e);
+		}
 	}
 }
